CHG: Implement tossEnemies()
* Decrease both popupTime and chainDelay so that the game gets harder.
* Increase the speed of our physics world, so that objects move rise and fall faster.
* Switch sequencePosition: Call createEnemy() correctly for the sequence type, passing in whether to force bomb creation or not.

Each sequence in our array creates one or more enemies, then waits for them to be destroyed before continuing.
Enemy chains are different: they create five enemies with a short break between, and don't wait for each one to be destroyed before continuing.

If nextSequenceQueued is false, it means we don't have a call to tossEnemies() in the pipeline waiting to execute. It gets set to true only in the gap between the previous sequence item finishing and tossEnemies() being called. Think of it as meaning, "I know there aren't any enemies right now, but more will come shortly."